Equidimensional Decomposition

Introduction

Functionality

Equidimensional Decomposition

Introduction

AlgebraicSolving.jl allows to compute equidimensional decompositions of polynomial ideals with defined over prime fields of characteristic smaller than $2^{31}$ and over the rationals.

More precisely, given a polynomial ideal $I$ it computes locally closed sets $X_1, \dots, X_k$ s.t. $V(I)=\bigcup_{i=1}^{k} X_j$ and such that each $X_j$ is equidimensional in the sense that the Zariski closure of each $X_j$ is equidimensional.

The implemented algorithm is the one given in this paper.

Functionality

Each locally closed set in the output of the implemented decomposition algorithm is of the form $V(F) \setminus V(g_1 \cdot \dots \cdot g_r)$ for a finite set of polynomials $F$ and polynomials $g_1,\dots, g_r$.

AlgebraicSolving.equations — Method

equations(X::LocallyClosedSet)

Given a locally closed set X of the form $V(F) \ V(g_1 \cdot \dots \cdot g_r)$, return the list of polynomials $F$.

AlgebraicSolving.inequations — Method

inequations(X::LocallyClosedSet)

Given a locally closed set X of the form $V(F) \ V(g_1 \cdot \dots \cdot g_r)$, return the list of polynomials $g_1, \dots g_r$.

AlgebraicSolving.Ideal — Method

Ideal(X::LocallyClosedSet)

Return a polynomial ideal I whose zero locus coincides with the Zariski closure of X.

Note: If no ideal I is known, this will require a Gröbner basis computation. The output of this function is cached.

AlgebraicSolving.equidimensional_decomposition — Method

equidimensional_decomposition(I::Ideal{T}, info_level::Int=0) where {T <: MPolyRingElem}

Given a polynomial ideal I, return a list of locally closed sets dec s.t. each locally closed set in dec is equidimensional (i.e. its Zariski closure has irreducible components of only one dimension) and s.t. the algebraic set defined by I equals the union of the locally closed sets in dec.

When the ground field of I is the rational numbers, a multi-modular method is used.

Arguments

I::Ideal{T} where T <: MPolyRingElem: input ideal.

info_level::Int=0: info level printout: off (0, default), details about multi-modular computation (1, only relevant when I is defined over the rational numbers), computational details (2)

Example

julia> using AlgebraicSolving

julia> R, (x, y, z) = polynomial_ring(GF(65521), ["x", "y", "z"])
(Multivariate polynomial ring in 3 variables over GF(65521), FqMPolyRingElem[x, y, z])

julia> I = Ideal([x*y - x*z, x*z^2 - x*z, x^2*z - x*z])
FqMPolyRingElem[x*y + 65520*x*z, x*z^2 + 65520*x*z, x^2*z + 65520*x*z]

julia> equidimensional_decomposition(I)
3-element Vector{AlgebraicSolving.LocallyClosedSet{FqMPolyRingElem}}:
 V(x*y + 65520*x*z) \ V(y + 65520*z)
 V(y + 65520*z, x*z^2 + 65520*x*z) \ V(x * (z + 65520))
 V(y + 65520*z, x*z^2 + 65520*x*z, x^2*z + 65520*x*z, z + 65520) \ V(x * z)