Krull dimension of an ideal

Introduction

Functionality

Krull dimension of an ideal

Introduction

AlgebraicSolving allows to compute the Krull dimension for the ideal spanned by given input generators over finite fields of characteristic smaller $2^{31}$ and over the rationals.

The underlying engine is provided by msolve.

Functionality

AlgebraicSolving.dimension — Method

dimension(I::Ideal{T}) where T <: MPolyRingElem

Compute the Krull dimension of a given polynomial ideal I.

Note: This requires a Gröbner basis of I, which is computed internally if not already known.

Examples

julia> using AlgebraicSolving

julia> R, (x, y, z) = polynomial_ring(QQ, ["x", "y", "z"]);

julia> I = Ideal([x*y,x*z,y*z]);

julia> dimension(I)
1