Real Algebraic Geometry

Introduction

Functionality

Real Algebraic Geometry

Introduction

AlgebraicSolving allows to work with real semi-algebraic sets, i.e. subsets of real affine space defined by polynomial equations and inequalities.

Functionality

AlgebraicSolving implements finding sample points per connected components of semi-algebraic sets:

AlgebraicSolving.points_per_components — Method

points_per_components(eqs::Vector{QQMPolyRingElem}, pos::Vector{QQMPolyRingElem}, ineqs::Vector{QQMPolyRingElem}; nr_thrds::Int=1, worker_pool::AbstractWorkerPool=default_worker_pool(), info_level::Int=0)::Vector{Vector{Vector{QQFieldElem}}}

Given a semi-algebraic set defined by equations eqs, positivity constraints pos, and non-vanishing constraints ineqs, returns a vector of points meeting all connected components of the set. Each point is represented by an isolating box, i.e., a vector of intervals represented by pairs of rational numbers.

NOTE: Sampling points per components is currently only implemented for systems without equations, i.e., eqs must be empty.

Arguments

eqs::Vector{QQMPolyRingElem}: equations defining the semi-algebraic set.

pos::Vector{QQMPolyRingElem}: positivity constraints defining the semi-algebraic set.

ineqs::Vector{QQMPolyRingElem}: non-vanishing constraints defining the semi-algebraic set.

nr_thrds::Int=1: the number of threads to use for parallel computations.

worker_pool::AbstractWorkerPool=default_worker_pool(): the worker pool to use for parallel computations.

info_level::Int=0: info level printout: off (0, default), summary (1), detailed (2).

Examples

julia> using AlgebraicSolving

julia> R, (x, y) = polynomial_ring(QQ, ["x", "y"])
(Multivariate polynomial ring in 2 variables over QQ, QQMPolyRingElem[x, y])

julia> length(points_per_components(QQMPolyRingElem[], [x, y], [x + y - 1])) >= 2
true

AlgebraicSolving allows to compute a roadmap for the real trace of the zero-set of the ideal spanned by given input generators over the rationals.

It assumes that the underlying algebraic set is smooth, and its real trace is bounded.

The underlying engine is provided by msolve.

AlgebraicSolving.roadmap — Method

roadmap(I::Ideal{T} where T <: MPolyRingElem, <keyword arguments>)

Given a radical ideal I with solution set X, that is smooth and whose real trace XR is bounded, return a roadmap of XR

The output is given as a Roadmap structure, encoding the recursive structure of roadmaps. It is encoded as a chained list, whose root is containing the equations defining X and each node representing a curve component, that is defined by additional polar equation and base point. Moreover it is linked to fibers, that share the same base point.

Arguments

I::Ideal{T} where T <: QQMPolyRingElem: input generators.

C::Vector{Vector{QQFieldElem}}=Vector{QQFieldElem}[]: query points with rational coefficients

info_level::Int=0: verbosity level

)

Examples

julia> using AlgebraicSolving

julia> R,(x1,x2,x3,x4) = polynomial_ring(QQ, ["x1","x2","x3","x4"])
(Multivariate polynomial ring in 4 variables over QQ, QQMPolyRingElem[x1, x2, x3, x4])

julia> I = Ideal([(x1^2+x2^2+x3^2+x4^2+9-1)^2-4*9*(x1^2+x2^2+x3^2) + 1])
QQMPolyRingElem[x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65]

julia> RM = roadmap(I)
Vector{Tuple{QQMPolyRingElem, QQFieldElem}}[[], [(x1, -3)], [(x1, -3), (x2, -2)], [(x1, -2)], [(x1, -2), (x2, -1)], [(x1, -2), (x2, 0)], [(x1, -2), (x2, 1)], [(x1, 3)], [(x1, 3), (x2, 2)]]

julia> nb_nodes(RM)
9

julia> all_eqs(RM)
9-element Vector{Ideal{QQMPolyRingElem}}:
 QQMPolyRingElem[x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65, -4*x1^2*x3 - 4*x2^2*x3 - 4*x3^3 - 4*x3*x4^2 + 40*x3, -4*x1^2*x4 - 4*x2^2*x4 - 4*x3^2*x4 - 4*x4^3 - 32*x4, 0, 0]
 QQMPolyRingElem[x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65, -4*x1^2*x4 - 4*x2^2*x4 - 4*x3^2*x4 - 4*x4^3 - 32*x4, x1 + 3]
 QQMPolyRingElem[x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65, x1 + 3, x2 + 2]
 QQMPolyRingElem[x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65, -4*x1^2*x4 - 4*x2^2*x4 - 4*x3^2*x4 - 4*x4^3 - 32*x4, x1 + 2]
 QQMPolyRingElem[x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65, x1 + 2, x2 + 1]
 QQMPolyRingElem[x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65, x1 + 2, x2]
 QQMPolyRingElem[x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65, x1 + 2, x2 - 1]
 QQMPolyRingElem[x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65, -4*x1^2*x4 - 4*x2^2*x4 - 4*x3^2*x4 - 4*x4^3 - 32*x4, x1 - 3]
 QQMPolyRingElem[x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65, x1 - 3, x2 - 2]

In addition, AlgebraicSolving can compute equations definition critical loci of polynomial maps over the given algebraic set.

AlgebraicSolving.computepolar — Method

computepolar(J::Union{Vector{Int},UnitRange{Int}}, V::Ideal{P} where P <: MPolyRingElem; <keyword arguments>)

Compute the polar variety associated with the map whose components are [phi_1, …, phi_p, x_{p+1}, …, x_n] indexed by J, for an algebraic variety defined by the ideal V.

More precisely, this function computes the set of points x in V such that, if psi denotes the above map, then the image of the tangent space Tₓ(V) under the differential psi has dimension strictly less than dimproj.

This is a key geometric construction in the computation of polar varieties, used in the roadmap algorithm and critical point methods.

Arguments

J::Union{Vector{Int},UnitRange{Int}}: Indices of the selected coordinate functions

V::Ideal{P} where P <: MPolyRingElem: Input ideal defining the variety V on which critical loci is computed

phi::Vector{P}=P[]: Polynomial map possibly completed by projection coordinates to have a map of total length n.

dimproj::Int=length(J)-1: Expected maximal dimension of the image of the tangent space. Typically equals the number of projection coordinates minus one.

only_mins::Bool=false: If true, only the computed minors of the Jacobian are returned; otherwise, the output includes both the generators of V and those minors.

Returns

If only_mins=false (default): A Vector{P} containing the union of the generators of V and the computed minors. This defines the ideal of the polar variety.

If only_mins=true: A Vector{P} containing only the minors (without the original equations of V).

Example

julia> using AlgebraicSolving

julia> R,(x1,x2,x3,x4) = polynomial_ring(QQ, ["x1","x2","x3","x4"])
(Multivariate polynomial ring in 4 variables over QQ, QQMPolyRingElem[x1, x2, x3, x4])

julia> I = Ideal([(x1^2+x2^2+x3^2+x4^2+9-1)^2-4*9*(x1^2+x2^2+x3^2) + 1])
QQMPolyRingElem[x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65]

julia> computepolar(1:3, I, dimproj=1, phi=[x1^2+x2^2+x3^2+x4^2])
5-element Vector{QQMPolyRingElem}:
 x1^4 + 2*x1^2*x2^2 + 2*x1^2*x3^2 + 2*x1^2*x4^2 - 20*x1^2 + x2^4 + 2*x2^2*x3^2 + 2*x2^2*x4^2 - 20*x2^2 + x3^4 + 2*x3^2*x4^2 - 20*x3^2 + x4^4 + 16*x4^2 + 65
 -4*x1^3 - 4*x1*x2^2 - 4*x1*x3^2 - 4*x1*x4^2 + 40*x1
 -4*x1^2*x4 - 4*x2^2*x4 - 4*x3^2*x4 - 4*x4^3 - 32*x4
 -2*x1
 -2*x4

AlgebraicSolving also implements functionality for real root classification:

AlgebraicSolving.real_root_classification — Method

real_root_classification(I::ParametricIdeal{K}, g::Vector{MPoly{K}}; <keyword arguments>) -> SemialgebraicSet

Computes the real root classification of the polynomials g with respect to the parametric ideal I. The result is a semi-algebraic set that describes the regions in the parameter space where the number of real roots such that g is positive is constant. Each region is accompanied by the constant number of real roots in that region, and a witness point in that region.

Note: The regions described by the semi-algebraic set may differ from the actual regions due to the fact that the Hermite matrices may not specialize well at certain parameter values. However, their symmetric difference is guaranteed to be contained in a proper algebraic set.

Arguments

I::ParametricIdeal{K}: input parametric ideal.

g::Vector{MPoly{K}}: input list of polynomials.

nr_thrds::Int=1: the number of threads to use for parallel computations.

worker_pool::AbstractWorkerPool=default_worker_pool(): the worker pool to use for parallel computations.

info_level::Int=0: info level printout: off (0, default), summary (1), detailed (2).

show_progress::Bool=info_level >= 1: whether to show progress bars during computations.

output_form::Symbol=:sign: the output form of the real root classification, can be either polynomials accompanied by their signs (:sign), Hermite matrices accompanied by their signatures (:signature), or Hermite matrices and multiplication matrices (:matrix).

ignore_no_real_roots::Bool=false: whether to ignore sign conditions with no real roots.

This is based on the computation of Hermite matrices:

AlgebraicSolving.hermite_matrix — Method

hermite_matrix(gb::Vector{T}, b::Vector{T}, g::T) -> MatElem

Computes the Hermite matrix associated with the polynomial g with respect to the Groebner basis gb and the monomial basis b.

AlgebraicSolving.hermite_matrix — Method

hermite_matrix(I::ParametricIdeal{K}, g::MPoly{K}; <keyword arguments>) -> MatSpaceElem{K}

Computes the parametric Hermite matrix associated with the polynomial g with respect to the parametric ideal I. The Hermite matrix is computed by evaluating the ideal at a number of parameter values and interpolating the coefficients of the resulting Hermite matrices.

Arguments

retry::Int=10: the maximum number of consecutive failures allowed when computing the Hermite matrix or interpolating the coefficients.

nr_thrds::Int=1: the number of threads to use for parallel computations.

worker_pool::AbstractWorkerPool=default_worker_pool(): the worker pool to use for parallel computations.

show_progress::Bool=false: whether to show a progress bar while computing the Hermite matrix.

AlgebraicSolving.hermite_matrix — Method

hermite_matrix(g::Vector{MPoly{K}}, α::Vector{Int}, I::ParametricIdeal{K}; <keyword arguments>) -> MatSpaceElem{K}

Computes the parametric Hermite matrix associated with the polynomials g raised to the powers specified in α with respect to the parametric ideal I. The Hermite matrix is computed by evaluating the ideal at a number of parameter values and interpolating the coefficients of the resulting Hermite matrices.

Arguments

nr_thrds::Int=1: the number of threads to use for parallel computations.

worker_pool::AbstractWorkerPool=default_worker_pool(): the worker pool to use for parallel computations.

show_progress::Bool=false: whether to show a progress bar while computing the Hermite matrix.

AlgebraicSolving.hermite_matrix — Method

hermite_matrix(g::Vector{MPoly{K}}, α::Vector{Int}, I::ParametricIdeal{K}, vals::Vector{QQFieldElem}; <keyword arguments>) -> QQMatrix

Computes the Hermite matrix associated with the polynomials g raised to the powers specified in α with respect to the parametric ideal I evaluated at the parameter values specified in vals. The Hermite matrix is computed by evaluating the ideal at the specified parameter values and interpolating the coefficients of the resulting Hermite matrices.

Arguments

nr_thrds::Int=1: the number of threads to use for parallel computations.

worker_pool::AbstractWorkerPool=default_worker_pool(): the worker pool to use for parallel computations.

show_progress::Bool=false: whether to show a progress bar while computing the Hermite matrix.